--- core/js/tsconfig.json
@@ -1,15 +1,0 @@
-{
-  "compilerOptions": {
-    "declaration": true,
-    "lib": ["es2019", "dom"],
-    "target": "ES2018",
-    "strict": true,
-    "moduleResolution": "node",
-    "baseUrl": ".",
-    "paths": {
-      "@lib/*": ["src/*"],
-    }
-  },
-  "include": ["."],
-  "exclude": ["node_modules/**", "**/dist/**"]
-}
