--- core/js/tsup.config.ts
@@ -1,11 +1,0 @@
-import { defineConfig } from "tsup";
-
-export default defineConfig([
-  {
-    entry: ["src/index.ts"],
-    format: ["cjs", "esm"],
-    outDir: "dist",
-    external: ["zod"],
-    dts: true,
-  },
-]);
