--- core/js/vitest.config.js
@@ -1,6 +1,0 @@
-import tsconfigPaths from "vite-tsconfig-paths";
-
-const config = {
-  plugins: [tsconfigPaths()],
-};
-export default config;
