+++ js/package.json
@@ -38,11 +38,18 @@
       "import": "./dev/dist/index.mjs",
       "module": "./dev/dist/index.mjs",
       "require": "./dev/dist/index.js"
+    },
+    "./util": {
+      "types": "./util/dist/index.d.ts",
+      "import": "./util/dist/index.mjs",
+      "module": "./util/dist/index.mjs",
+      "require": "./util/dist/index.js"
     }
   },
   "files": [
     "dist/**/*",
-    "dev/dist/**/*"
+    "dev/dist/**/*",
+    "util/dist/**/*"
   ],
   "scripts": {
     "build": "cross-env NODE_OPTIONS=\"--max-old-space-size=8192\" tsup",
@@ -97,7 +104,6 @@
   },
   "dependencies": {
     "@ai-sdk/provider": "^1.1.3",
-    "@braintrust/core": "workspace:*",
     "@next/env": "^14.2.3",
     "@vercel/functions": "^1.0.2",
     "argparse": "^2.0.1",
