+++ js/util/bytes.ts
