+++ js/util/graph_util.ts
