+++ js/util/merge_row_batch.ts
