+++ js/util/score.ts
