+++ js/util/span_identifier_v2.ts
