+++ js/util/span_identifier_v3.ts
