+++ js/util/string_util.test.ts
