+++ js/util/string_util.ts
