+++ js/util/type_util.ts
